requirements.txt
pywin32>=305
openai>=1.0.0
pandas>=2.0.0
networkx>=3.0
tk>=0.1
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.0.0
